Removed menu submodule and finallised the TimeoffMenuCommand execution structure.
