--- src/app/timeoff/menu/mod.rs
@@ -1,2 +1,0 @@
-mod command;
-pub(crate) use command::*;
